// src/app/shared/interfaces.ts
export type TsCallingPageType = 'list' | 'create' | 'edit' | 'review';
export interface TsCallingPage {
  callingPage: TsCallingPageType;
}


export interface TsMapRequest {
  command: 'replace' | 'add' | 'rem' | 'clear';
  pathId: string;
  plotType: 'active' | 'overlay';
  colour: string;
}
// ...
export interface TsFeatureCollection {
  type: 'FeatureCollection';
  features: TsFeature[];
  bbox?: TsBoundingBox;
  properties?: TsGeoJsonProperties;
}

export interface TsFeature {
  bbox?: TsBoundingBox;
  id?: string; // mapbox
  type: 'Feature';
  geometry: TsGeometry;
  // properties: TsGeoJsonProperties | TsMapboxProperties | null;
  properties: TsGeoJsonProperties | null;
}

export type TsGeometry = TsPoint | TsLineString;

export interface TsPoint {
  type: 'Point';
  coordinates: TsPosition;
}

export interface TsLineString {
  type: 'LineString';
  coordinates: TsPosition[];
}

export type TsPosition = [number, number];

/*minLng, minLat, maxLng, maxLat*/
export type TsBoundingBox = [number, number, number, number];
// export interface TsBoundingBoxObject {
//   minLat: number;
//   maxLat: number;
//   minLng: number;
//   maxLng: number;
// }

export interface TsGeoJsonProperties {
    pathId?: string;
    info?: TsInfo;
    params?: TsParams;
    stats?: TsPathStats;
    lineColour?: string;
    creationDate?: string;
    lastEditDate?: string;
    plotType?: string;
    userID?: string;
    latitude?: number;   // front need, used by map service when creating points geoJson
    matched?: boolean;
    title?: string;

}

export interface TsMapboxProperties {
  description?: string;
  title?: string;
} // mapbox labels}

export interface TsInfo {
    direction: string;
    category: string;
    nationalTrail: boolean;
    name: string;
    description: string;
    pathType: string;           // 'route' or 'track'
    startTime: string;
    isLong: boolean;
    isPublic: boolean;
    createdBy: string;
    isElevations: boolean;
    activityType: TsActivityType;
    startPoint: TsPosition;
}

export interface TsParams {
    elevs: number[];
    time: number[];
    heartRate: number[];
    cadence: number[];
    cumDistance: number[];
    matchedPoints?: number[][]; /// used for debugging route algorthims on the back end
}

export interface TsPathStats {
    bbox: {
        minLng: number,
        minLat: number,
        maxLng: number,
        maxLat: number
    };
    nPoints: number;
    simplificationRatio: number;
    duration: number;
    distance: number;
    pace:  number;
    elevations: {
        ascent: number,
        descent: number,
        maxElev: number,
        minElev: number,
        lumpiness: number,
        distance: number,
        nPoints: number,
        badElevData?: boolean
    };
    p2p: {
        max: number,
        ave: number
    };
    movingStats: {
        movingTime: number,
        movingDist: number,
        movingPace: number,
    };
    hills: TsHill[];
    splits: {
        kmSplits: number[][],
        mileSplits: number[][]
    };
}